% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelateMotifswithGenes.R
\docType{methods}
\name{filterActivityByTFexpression}
\alias{filterActivityByTFexpression}
\title{# Filter motifs in activity matrix by their correlation with their TF expression}
\usage{
filterActivityByTFexpression(
  coef,
  gene.expression,
  threshold = 0.6,
  method = "spearman",
  cores = 1,
  abs = FALSE,
  split.sign = NULL
)
}
\arguments{
\item{coef}{a numeric matrix indicating an activity motif matrix}

\item{gene.expression}{a numeric matrix indicating a gene expression matrix
that is used to correlate with a `coef` entries}

\item{threshold}{a numeric value indicating correlation threshold}

\item{method}{a character indicating a method used for correlation.
By default "pearson", for more details check out
the `method` argument of the `stats:cor` function.}

\item{cores}{a numeric indicating number of cores. Default: 1.}

\item{abs}{a boolean value indicating whether calculate absolute value of
correlation. Defalt: FALSE.}

\item{split.sign}{a character value}
}
\value{
returns a \code{gtable} object from a ggplot R package
}
\description{
# Filter motifs in activity matrix by their correlation with their TF expression
}
