% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalMatrix.R
\docType{methods}
\name{filterSignalMatrix}
\alias{filterSignalMatrix}
\title{# Get a score in each window}
\usage{
filterSignalMatrix(
  SignalMatrix,
  signal.granges,
  gene.matrix,
  abs.corr.threshold = 0.5,
  corr.method = "spearman",
  species = "hg38",
  rule = "basalPlusExt",
  plot = FALSE,
  cores = 1
)
}
\arguments{
\item{target}{a path to a tabix file}

\item{windows}{a GRanges to be overlapped with ranges in target}

\item{strand.aware}{if TRUE (default: FALSE), the strands of the windows
will be taken into account. If the strand of a window is -,
the values of the bins for that window will be reversed}

\item{save.db}{a logical to decide whether regional counts within given windows
should be saved as flat file database or not}

\item{suffix}{a character string to append to the name of the output
flat file database, only used if save.db is true}

\item{cov.bases}{number minimum bases covered per region (Default:0).
Only regions with base coverage above this threshold are returned.}

\item{chunk.size}{Number of rows to be taken as a chunk for processing the methylDB
objects (default: 1e6)}
}
\value{
returns a list of \code{SignalMatrix} object that is
		   a numeric matrix of scores such as DNA methylation values
		   in each target region (rows) in each sample (columns) and
		   a GRanges object
}
\description{
# Get a score in each window
}
\examples{

library(methylKit)
file.list=list( system.file("extdata", "test1.myCpG.txt", package = "methylKit"),
               system.file("extdata", "test2.myCpG.txt", package = "methylKit"),
               system.file("extdata", "control1.myCpG.txt", package = "methylKit"),
               system.file("extdata", "control2.myCpG.txt", package = "methylKit") )
myobj=methRead( file.list,
           sample.id=list("test1","test2","ctrl1","ctrl2"),assembly="hg18",
           pipeline="amp",treatment=c(1,1,0,0))
methylbase.obj=unite(myobj)
methylbase.objDB=makeMethylDB(methylbase.obj, dbdir = "methylDB")

windows=GRanges("chr21",
IRanges(start=c(9853296,9906604), end=c(9860126,9906616)))


sig.mtx=getSignalMatrix(
             Rsamtools::TabixFile(methylbase.objDB@dbpath),
            windows,
            methylbase.objDB@sample.ids,
            methylbase.objDB@assembly,
            methylbase.objDB@treatment)
            
gene.matrix =  matrix(10:17, 2,4)   

fil.gene.sign.mtx = 
   filterSignalMatrix(sig.mtx,
                      windows,
                      gene.matrix,
                      corr.method="spearman",
                      species="hg38",
                      rule="basalPlusExt",
                      plot = FALSE,
                      cores=1)        

}
