% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{methods}
\name{getActiveGenes}
\alias{getActiveGenes}
\title{# Perform differential gene expression analysis using DESeq2 in order to obtain
# active/expressed genes.}
\usage{
getActiveGenes(
  gene.expression,
  conditions,
  case,
  control,
  covariates = NULL,
  rowsums = 1,
  padj = 0.1,
  log2foldchange = 1
)
}
\arguments{
\item{gene.expression}{a numeric matrix indicating gene expression.
In columns are samples and in rows genes.}

\item{conditions}{a data.frame with a one column `condition` and row names
indicate names of samples the same as in column names
of the `gene.expression` matrix}

\item{covariates}{a data.frame that contains covariates for differential
expression analysis. Each column indicates a covariate
and in rows are samples.}
}
\value{
returns
}
\description{
It requires minimum of 2 biological replicates for each condition.
}
\examples{
# TODO

}
