% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{methods}
\name{getModel}
\alias{getModel}
\title{# Get a model}
\usage{
getModel(signal, motifhits, cores = 1, transform.meth = FALSE)
}
\arguments{
\item{signal}{a SignalMatrix object or a numeric matrix}

\item{motifhits}{a MotifHitsMatrix object or a numeric  matrix}

\item{cores}{a number of cores}

\item{transform.meth}{a boolean whether log transform input \code{signal}
using a following transformation: log2((Y+1)/(100-Y+1))}
}
\value{
returns a \code{Model} object
}
\description{
# Get a model
}
\details{
https://journal.r-project.org/archive/2012-2/RJournal_2012-2_Kloke+McKean.pdf
}
\examples{

signalmatrix_example_path = system.file("extdata", 
                                   "signalmatrix_example.txt", 
                                   package = "motifActivity")
signalmatrix_example = read.table(signalmatrix_example_path, header=TRUE)

motifhitsmatrix_example_path = system.file("extdata", 
                                       "motifhitsmatrix_example.txt", 
                                       package = "motifActivity")
motifhitsmatrix_example = read.table(motifhitsmatrix_example_path, header=TRUE)

model=getModel(signalmatrix_example,
              motifhitsmatrix_example,
              cores=1,
              transform.meth=TRUE)

}
