% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifHitsMatrix.R
\docType{methods}
\name{getMotifsHits}
\alias{getMotifsHits}
\title{# Calculate number of occurences of DNA motifs in given sequences}
\usage{
getMotifsHits(
  pfms,
  signal,
  bg,
  order = 1,
  normalize = TRUE,
  singlestranded = FALSE,
  cores = 1,
  genome = NULL,
  ...
)
}
\arguments{
\item{pfms}{a list of PWM matrices}

\item{signal}{genomic regions either as a GRanges object.
If a GRanges if provided then genome argument is required.}

\item{bg}{a BSgenome object}

\item{order}{order of the Markov models that will be used as the background model.
Default: order = 1}

\item{normalize}{normalizes a PFM and optionally adds pseudo-evidence to each entry
of the matrix. It internally uses the motifcounter::normalizeMotif
function.}

\item{cores}{a numeric indicating number of cores. Default: 1.}

\item{genome}{a BSgenome object}

\item{...}{parameters to pass to the motifcounter::motifcounterOptions function}
}
\value{
returns a \code{MotifsHitsMatrix} object that is
		   a matrix of occurrences DNA motifs in given genomic regions
}
\description{
# Calculate number of occurences of DNA motifs in given sequences
}
\examples{

library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
genes=genes(txdb)
genes=keepSeqlevels(genes, "chr12", pruning.mode="coarse")[1:50]
library(RCAS)
bg=createControlRegions(genes)

DNAmotifs.pfms <- as.list(query(query(MotifDb, "Hsapiens"),
                                "jaspar2018"))[1:10]
library(BSgenome.Hsapiens.UCSC.hg38)
genome=Hsapiens

# run finding motif hits in given windows
motifhits = getMotifsHits(DNAmotifs.pfms,
                          genes,
                          bg,
                          normalize=TRUE,
                          singlestranded=FALSE,
                          cores=1,
                          genome=genome,
                          order=0,
                          alpha=0.0001, gran=0.05)


}
