% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifHitsMatrix.R
\name{getSequences}
\alias{getSequences}
\title{# Get nucleotide sequences from a given windows}
\usage{
getSequences(windows, genome, background = FALSE, order = 1, cores = 1)
}
\arguments{
\item{windows}{a GRanges object}

\item{genome}{a BSgenome object}
}
\description{
# Get nucleotide sequences from a given windows
}
