% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifActivity-classes.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{An S4 class for storing \code{getModel} function results}
\description{
The resulting object stores the results of \code{getModel} function
}
\section{Slots}{
\describe{
                 \item{\code{signal}: XYZ},
         \item{\code{motif.hits}: XYZ}
                }
}

\seealso{
\code{\link{getModel}}
}
