% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifActivity-classes.R
\docType{class}
\name{MotifHitsMatrix-class}
\alias{MotifHitsMatrix-class}
\title{An S4 class for storing \code{getMotifHits} function results}
\description{
The resulting object stores the results of \code{getMotifHits} function
}
\section{Slots}{
\describe{
                 \item{\code{matrix}: XYZ}
                }
}

\seealso{
\code{\link{getMotifHits}}
}
