% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualise.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\title{# Plot a heatmap of motifs activities}
\usage{
plotHeatmap(
  values,
  col,
  top.annotation = NULL,
  row.annotation = NULL,
  col.top.annotation = NULL,
  col.row.annotation = NULL,
  title = "Heatmap",
  cluster_columns = TRUE,
  match.heat.dend = FALSE
)
}
\arguments{
\item{values}{a numeric matrix}

\item{top.annotation}{a list of vectors to plot as a top annotation}

\item{row.annotation}{a list of vectors to plot as a row annotation}

\item{col.top.annotation}{a list of colors that correspond to `top.annotation`}

\item{col.row.annotation}{a list of colors that correspond to `row.annotation`}
}
\value{

}
\description{
# Plot a heatmap of motifs activities
}
