% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualise.R
\docType{methods}
\name{plotMotifActivity}
\alias{plotMotifActivity}
\title{# Plot motif activity for a given DNA motif}
\usage{
plotMotifActivity(
  motif.name,
  coef,
  title = motif.name,
  colour.line = "black",
  colour.xaxis = "black",
  ylab = "Motif activity",
  match.clustering = TRUE,
  legend.position = "top"
)
}
\arguments{
\item{motif.name}{a character indicating a name of a DNA motif}

\item{coef}{a matrix of coefficients. In columns are samples, in rows motifs.}

\item{title}{a string indicating a title of the plot. By default it's `motif.name`.}

\item{colour.line}{a character indicating a color of a line}

\item{colour.xaxis}{a character indicating a color of labels on X axis}

\item{legend.position}{a character indicating a position of legend, such as 
"left", "top", "right", or "bottom".}
}
\value{
returns a \code{gg} object from a ggplot R package
}
\description{
# Plot motif activity for a given DNA motif
}
