% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualise.R
\docType{methods}
\name{plotNetwork}
\alias{plotNetwork}
\title{# Plot transcription factors in a PPI network build from the STRING database}
\usage{
plotNetwork(
  genes.hgnc,
  species = 9606,
  version = "11",
  score_threshold = 200,
  input_directory = "./",
  nodes.color = NULL,
  mark.color = NULL,
  cluster.nodes = TRUE,
  plot_dendrogram = FALSE,
  cut_at = NULL,
  add.missing = FALSE,
  graph_sub = NULL,
  ...
)
}
\arguments{
\item{genes.hgnc}{a character vector indicating names of genes in the HUGO Gene Nomenclature}

\item{species}{a number corresponding to a species required for the STRINGdb$new function}

\item{version}{a number indicating a version of the STRINGdb database}

\item{score_threshold}{a number indicating a STRING combined score threshold 
(the network loaded contains only interactions having a combined 
score greater than this threshold)}

\item{nodes.color}{a character vector indicating color of nodes}

\item{mark.color}{a character vector indicating color of nodes}

\item{cluster.nodes}{a boolean value indicating whether to cluster nodes and color-code them
together}

\item{plot_dendrogram}{a boolean value indicating whether to plot dendogram of clustered
nodes}

\item{cut_at}{a numeric value indicating a cut off for number of nodes}

\item{graph_sub}{a graph object from the igraph R package indicating graph of TFs}
}
\value{

}
\description{
# Plot transcription factors in a PPI network build from the STRING database
}
