% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifActivity-classes.R
\docType{class}
\name{SignalMatrix-class}
\alias{SignalMatrix-class}
\title{An S4 class for storing \code{getSignal} function results}
\description{
The resulting object stores the results of \code{getSignal} function
}
\section{Slots}{
\describe{
                 \item{\code{matrix}: XYZ},
         \item{\code{signal.type}: XYZ}
                }
}

\seealso{
\code{\link{getSignalMatrix}}
}
